let Emacs adjust point
Previously we did adjust point ourselves, because "Emacs often leaves
point in invisible regions, it seems."  However we did not actually
succeed in doing so because `disable-point-adjustment' is only set to
nil when a Magit mode is turned on, not for each and every command that
could possibly be run in a Magit buffer.  But that variable's doc-string
reveals that this doesn't have the desired effect:

  After a command is executed, if point is moved into a region that has
  special properties (e.g. composition, display), we adjust point to the
  boundary of the region.  But, when a command sets this variable to
  non-nil, we suppress the point adjustment.

  This variable is set to nil before reading a command, and is checked
  just after executing the command.

So I am removing our code that does the same, doing so does indeed not
change the behavior.