magit-init: use magit-run-git
In `magit-init' use `magit-run-git' instead of the more low-level
`magit-run-git*'.

The latter should only be used in special cases, and should otherwise be
avoided.  Creating a repository now results in a refresh which is rather
pointless and has no effect because there cannot be any buffers to
refresh yet.  On the other hand checking whether anything has to be
refreshed also isn't costly enough to actively avoid it in this case.
The benefit is that I will not once more have to think about whether
there actually a reason `magit-run-git*' is used :-)