magit-run-git-with-input: fake the synchronous process here
Only asynchronous processes can receive input.  So even if we want to
wait until the subprocess has finished we use an asynchronous process,
and then just loop until it is done.

Previously this was done in `magit-run*' but that caused that do be much
more complicated than it need be.  `magit-run-git-with-input' is much
more suitable so we do the looping here now.  (Actually `magit-run*'
also still supports looping, but that will change in few commits.)