magit-{run-git-with-logfile,process-logging-filter}: new functions
Only asynchronous processes can have a filter which we need to be able
to log to a file and send the output to *magit-process* too.  So when
we also want to wait until the subprocess has finished we use an
asynchronous process, and then just loop until it is done.

Now there is a dedicated function to this sort of thing.  This will
allow simplifying `magit-run*'.