magit-run*: with input always use pipe
When sending input to a process always use a pipe (otherwise that
depends on option `magit-process-connection-type').  We have done
that ever since issue #20 was "fixed" but not in all cases.

Previously we did so when sending input to a "synchronous" process
(which is an asynchronous process, but we wait until it is done
before returning), but not for asynchronous processes.  That didn't
makes any sense.

This might, or might not, fix #487.  (I am just guessing, I don't
have Windows and until someone who does, actually investigates the
root cause of that issue, guessing is all I can do.  Let's hope I
am guessing right.)