magit-process-finish: take a single argument
For asynchronous process we are going to need to provide more
information to this function, which is actually contained within the
process object.  For synchronous process object exists, so we had the
callers provide each required information as a separate argument.

Simplify that my replacing these arguments with just one PROCESS
argument, which may also be a list, consisting of the values in the
correct order.