support multiple synchronous processes
That was easy.  Basically Magit had to be told not to refuse to start
a second process.  Well almost - I first had to make sure that process
output isn't clobbered when several processes write to the same buffer,
but maybe we should have taken care of this a bit earlier: close #24,
the oldest issue which was still open.