magit-{call,start}-process: flatten arguments ARGS
Flatten the first level of `magit-call-process' and
`magit-start-process's rest argument ARGS before passing the arguments
to `process-file' and `start-file-process'.  This also affects
`magit-call-git and `magit-start-git', as well as the high-level api
functions `magit-run-git{*}'.

This in most cases (all as of now) makes it unnecessary to use `apply'
when the arguments to git (or some other process) have to determined
at runtime.

Function`magit-run-git-async' shows the ARGS in the echo area, so
flatten them here too.