magit-process-buffer: properly setup buffer
Teach `magit-process-buffer' how to properly setup the buffer, i.e
create the root section when called with non-nil CREATE and if the
buffer doesn't exist yet.

Previously this was done in `magit-process-setup', but while that is
only suitable right before starting a process, we might actually need
to display the buffer before any process has run, as is currently the
case for `magit-git-command'.

Close #1138.