magit-push-dwim: extract from magit-push
Remove the default push implementation from the `magit-push' command,
leaving nothing but calling `magit-push-hook' behind.  Create a new
function `magit-push-dwim' from the removed code and add it to the
default value of the hook.  This allows using a simpler implementation
that is more true to Git, while not disturbing users who have grown
accustomed the, somewhat wacky, old implementation.