magit-log-reflog-re: deal with git-merge inconsistency
Other commands put the revs and refs they deal with after the colon.
Before the colon goes the command and in some cases details about the
command, i.e. what variant of the command is used or in what mode
command was performed.

  rebase -i (finish): commit message

`git-merge' confuses the order and the position of the colon.

  merge branch-name: Fast-forward

Adjust the regexp to pretend the colon comes before the branch-name.
Ignore the detail part for now.