magit-log-margin-set-timeunit-width: new function
Extract `magit-log-margin-set-timeunit-width' from
`magit-refresh-log-buffer' so that it can be used elsewhere to.

Actually we should refactor the involved variables so that this only
has to be done when setting the "main" variable.  But there are other
things that should be improved here.  For users sake delay addressing
this particular point to avoid changing the options several times.