magit-section-action: dumb down message about unsupported section
Previously we attempted to include information about the type of the
section which does not support a particular action in the warning
message.  As identifier the value of `magit-section-type', a symbol, was
used.  If the `magit-description' property was set for that symbol, then
that was used.  However there was not a single symbol for which this was
the case.  Even if we did set such a description for all relevant
symbols, that wouldn't do us much good.  Many sections are only properly
identified using a vector.  Since vectors aren't symbols, they don't
have symbol properties...