magit-section-{case,action}: change order of type context vectors
Expect clause type context vectors to be in the order in which they
are actually processed.  That is the most narrow section now has to
come first.  If that doesn't match, then the rest doesn't even matter.
If it matches go on checking whether the parent (second element) also
matches and so on, until no more ancestor requirements are defined.