finally put the TITLE slot of magit-section structs to rest
I have prepared for this a long time ago but have hesitated to make the
switch because at the time there were more code areas that required
changes back then.  Now that these messy areas have been cleaned up or
removed the switch can be done with more confidence.

Ever since I prepared for the removal of TITLE, that slot and INFO
contained the same information, except in one case which gets fixed in
this commit.