add top-level "Modes" section
The (now) subsections that contain the definitions of individual
modes are currently spread across `magit.el' with unrelated sections
in between.  So as of now there are actually two top-level sections
called "Modes (N)".  That will be rectified eventually.

One of the subsections of "Modes (2)" is "Ediff Support".  Yes that
is wrong.  That section does not define a `magit-mode' derived mode.
Instead it defines (rather poorly) wrappers around Ediff commands.
I haven't decided yet where I want this code to end up but since I
am going to write `magit-ediff-*' variants of most `magit-diff-*'
commands, on possible approach would be to define these things in
the same place.  For now let's life with this inconsistency.