add top-level "Settings" section and a eof section
This section contains all the things a user can tweak to adapt Magit to
her wishes.  I think it helps users to have such a "all the settings are
to be found here" section.

The new section absorbs the existing top-level sections "Options" and
"Keymaps".  Some of the subsections that already existed in these old
sections, have new names now which better reflect their new surrounding.

The definitions of two `:set' functions for custom options are moved
right before their first uses, as we shouldn't begin the "Settings"
section with code, that would kinda defeat the purpose.  Having them
later in between custom options also isn't optimal but since these
functions have to be defined before they are first used their aren't
all that many alternatives.

Move the calls to `define-obsolete-variable-alias' to the very end of
the file into a new section "magit.el ends soon" which also contains
things like providing the library feature and setting buffer local
variables.  And yes that's a comment on the library convents that in
the age of dvcs require files to end with "this is the end".  Also
I just couldn't think of a better name: "Provide feature and somewhat
related stuff and final hacks" doesn't seem so good :-)