add new top-level sections "Magit Api" and "Porcelain"
* Add new top-level section "Magit Api", turning the former top-level
  sections containing the "section", "process" and "mode" code into
  sub-sections.  Rename these subsections.

* Add new top-level section "Porcelain", turning the former top-level
  sections containing the "apply", "visit" and "act" code into
  sub-sections.  Keep the old names of the subsections - cannot think
  of good ones right now.

* Add new top-level section "(misplaced)", group some former top-level
  sections below.  Now there are fewer top-level sections that are
  misplaced.

This isn't the end.  There is no one right way to arrange stuff.  This
is an arrangement that is suitable now.  It contains inconsistencies
but it is sufficient.  The same order would not have worked some time
ago and future changes will make further adjustments necessary.

This fucks with blaming and bisecting.  I won't compromise because our
version control tools are insufficient.  I maintain such a tool.  So
if desirable changes lead to suffering, then I do them anyway and then
I improve the tool.