log: respect core.abbrev
Remove the variable `magit-sha1-abbrev-length' and instead just respect
the git variable `core.abbrev'.  This allows users to set this per
repository.

Also remove the function `magit-diff-abbrev-arg' which used the value
of the option and returned "--abbrev=<N>".  Beside that, the only place
where the elisp variable was used was when aligning lines in logs that
contained only part of the graph.  Because we move the hash to the
beginning of the line we have to align based on the default length of
the hash.  We can get that by using `git config core.abbrev' once per
log, so it's cheap.