magit-log-count: limit scope
We only cut of logs after a certain number of commits when they are
shown in a log buffer, other logs usually are not long enough for that
to be necessary anyway.  Nevertheless, we now establish the dynamically
bound counter in `magit-wash-log', which is used for all logs.  Doing
so is cheap because it only amounts to calling `cl-incf' N times, where
N is usually fairly low.