magit-key-mode-map: new variable
This keymap is used as the parent of all generated popup keymaps.
Having such a parent saves us from having to define the same bindings
over and over again, and might in some cases allow users to customize
the bindings of all popups (for things that don't get overridden by
the individual popup keymaps).