magit-key-mode.el: refactor button and popup handling
Remove the existing maze of drawing functions consisting of
`magit-key-mode-draw-switches', `magit-key-mode-draw-options',
`magit-key-mode-draw-actions', `magit-key-mode-draw-buttons', and
`magit-key-mode-draw-in-cols'; replacing them with just
`magit-popup-insert-buttons' and `magit-popup-format-button', and
doing a bit more work in `magit-refresh-popup-buffer'.

This change could theoretically come with a bunch of small improvements
to popup appearance and at the same time remove a few lines of code.
But I avoid doing so for now.

The new options are `magit-popup-switch-format',
`magit-popup-option-format', and `magit-popup-action-format'.