magit-invoke-popup-option: disable options properly
Previously providing no input would not disable the option (but visually
indicated that it did) but just set the value to the empty string.  Now
exiting the minibuffer with an empty string means to turn off the
option.  Exiting with a whitespace-only string means use the option but
with no value (i.e. the value is the empty string).