magit-key-mode.el: rename faces
Change the prefix of all popup faces from `magit-key-mode-' to
`magit-popup-', like we have already done for many other symbols,
and remove the redundant `-face' suffix (that should only be used
for variables that can be used to select a face).

Rename `magit-key-mode-BUTTON-face' to `magit-popup-KEY', because it is,
and always was, used for the key part *of* a button, not *the* button.

Rename `magit-key-mode-SWITCH-face' to `magit-popup-ARGUMENT', because
starting with the next commit we will also use it for options, the other
kind of arguments.

Rename `magit-key-mode-OPTION-face' to `magit-popup-OPTION-VALUE',
because it is, and always was, used for option *values* never for the
option itself.

Make *most* of the old faces alias for the new face and mark them
obsolete.

Do *not* define an alias for the renamed `magit-key-mode-args-face'; if
users and theme authors did take inspiration from the default look, then
we *want* to break that, as the text propertized with this face is not
editable the same way a `widget-field' is.  The actual face definition
still has to be adapted accordingly.