magit-popup-mode-{setup,display-buffer}: new functions
Define `magit-popup-mode' using `define-derived-mode' and move
most of the code it used to contain into the new functions
`magit-popup-mode-setup' and `magit-popup-mode-display-mode'.

The new functions are modeled and named after `magit-mode-setup'
and `magit-mode-display-mode' but are implemented independently,
because `magit-key-mode' should not depend on `magit' and because
we actually need these functions to behave differently.

The various `magit-key-mode-popup-{*}' function now have to call
`magit-popup-mode-setup' instead of `magit-popup-mode'.