magit-popup-buffer-name: remove variable
When defining a popup the full symbol name has to be specified, and that
symbol, which is supposed to properly prefixed with a package "namespace"
(currently only "magit-") and by convention end with "-popup", is now the
only popup identifier.  Additionally prefixing that with "*magit-key: "
in the buffer name is redundant.