magit-define-popup: new macro
This completes the move from one monolithic variable that holds the
data of all existing popups, to separate variables for each popup.
That makes it *much* easier to customize existing popups and add new
ones.

The intermediate step was, that to define a popup one had to manually
define a command, an alist variable, and a map variable.  Now all of
that can be done using this new macro.  By the way the map variable
will soon be made unnecessary and instead a new "default argument
custom option" will be added.  If it weren't for this macro such
changes would require many changes to boilerplate code all over the
place.