popups: new symbol names for magit popups
The popup command and variable are now both usually named
`magit-VERB-popup'. VERB should be in infinitiv form (e.g. "commit").

Previously the command was named `magit-key-mode-popup-VERB', and VERB
was in -ing form.  The verb form is (and was) just a convention, and it
didn't even have to be a verb.  Actually it is now possible to use any
symbol name, because now the full symbol name has to be given.

We do not have to enforce uniformity by constructing the symbol names,
as there are other ways of enforcing consistency within Magit, and if
someone should happen to re-use the popup functionality outside of
Magit, then they are probably happy these symbols no longer are prefixed
with `magit-'.

Actually there is one more variable per popup: `...-map'.  That will be
removed soon, but yet another additional variable per popup will be
added, which will be named `...-defaults'.  In both cases the suffix is
fixed.