popups: use event sequences of constant length
Give all event sequences in popups a constant length.  For switches
and options always use a sequence that is two characters long; and
for actions always use a sequence that is a single character long.

There were only two popups that used sequences that did not already
conform with this new restriction: `magit-log-popup' and
`magit-commit-popup'.  In both cases it was not necessary to use
longer sequences in any way, there are enough fine shorter sequences
left that are nicer to use.

One reason for this change is that these shorter key sequences are
nicer.  Another reason, which should not be concealed is that imposing
this restrictions will make many upcoming changes much easier.