popups: represent each event using a single characters
Since we only support event sequences that are one event long (for the
time being) use a character instead of a string to represent such an
event.  (Argument sequences are actually two events long, but the first
event is constant across each argument type and cannot be changed for an
individual popup definition or even for an individual popup event
definition.  That's also a temporary restriction.)