popups: add support for default arguments
Allow specifying arguments that are enabled by default when defining a
popup using `magit-define-popup'.  This feature used to be implemented
in a very hackish way, so it was removed earlier in this patch series.

Now that the feature is back, and better than ever, also restore the
defaults that used to exists.  Actually there was only one.

The next commit will add support for customizing these defaults.