magit-popup-setvar-mode: new mode which uses "flat options"
For each option a custom variable is defined which holds the default
arguments.  The format (an alist) of the values of these options is
optimized for use inside popups.

But for certain popups it is desirable to use the value of that option
outside of the popup.  In such cases the value should be a list of
arguments (strings) which can be passed to a process as command line
arguments, not a list of cons cells which represent such arguments but
cannot be used without being reformatted.

Using the new `magit-popup-setvar-mode' for some popup allows doing
just that.