magit-diff: improve doc-string
Yes, that really is an improvement - it's less misleading. "Diffing
a range" means to compare two endpoints, while log actually does
something with every commit in between those.

We still use `magit-read-rev-range' to get the endpoints, but that
has to be fixes as part of #876.  We keep using that function for
now because that allows using "a..b" OR "a...b", but that also isn't
obvious yet, unfortunately.