magit-diff-popup: new popup
This is the first popup which can be used to invoke git with some
arguments as well as to set the variable used when refreshing the
current buffer to these arguments.  It is therefor possible to either
create a new diff buffer using certain arguments, or to update the diffs
in the current buffer.  That buffer isn't necessarily a diff-only
buffer, the status buffer for example can also contain diffs.

For diff buffers the ability to set the arguments used for refreshing as
a buffer-local value isn't strictly necessary, one could just invoke the
command that created the buffer again, this time using different
arguments.  This isn't the case for the status buffer, there is no way
to pass arguments intended for `git diff' to `magit-status'.  In the
future we might do the same for `git log' arguments.