magit-log-{*}: make more intuitive
* Use the terms "Oneline" and "Verbose" to describe the two log types.
  Previously "Short" and "Long" was used, and the latter is rather
  ambitious and depending on how you look at it simply wrong.
* Make sure the action buttons are arranged in a helpful manner.
* Rename the log commands to better reflect their purpose.
* Move the heavy lifting to the "core" functions instead of
  offloading it to the "variations".

And most importantly focus on showing logs for refs and revs, not
ranges.  Showing logs for ranges really isn't all that useful, there are
cases when it is, but showing the history of a branch different from the
current branch is VERY useful.  That was previously already possible by
abusing the "ranged log" to show a branch (by omitting the end of the
range), but that was VERY confusing to users.  Now one has to abuse the
"branch log" to show a "ranged log" and in that case the completion is
less helpful than it used to be.  That will change though.