log: goto commit that was current in previous buffer
`magit-mode-refresh-buffer' which is used to refresh any magit
buffer (which includes creating it) tries to keep point at the
same section.  However it only knows about one buffer at a time.

So if point is e.g. at a commit inside one of the "mini logs" in the
status buffer and the log buffer is brought up, then point is usually at
the beginning of the buffer but may also end up at some random position.

This commit teaches log how to goto the commit that is at point in the
buffer that was current before the log buffer was created.  This is done
mostly for logs that are used to select a commit.  Support for such
special logs will be added in the next commit.  But instead of limiting
it to just these special logs we do this here for all logs.  This should
be further generalized so that it can be used in other situations too.