magit-commit-{fixup,squash}: default to commit at point
Previously these commands either used the commit at point (or the marked
commit), no questions asked, or they had the user make a choice without
providing a default.  Now a default is provided by putting point on the
commit that was at point in the previous buffer.

This also looses the ability to use the marked commit (whether
unconditionally or as default).  That option being available was the
result of me being urged to merge the squash functionality into master
before I considered it fully ready.