branch: use common "magit-branch-" prefix
Rename `magit-create-branch' to `magit-branch-and-checkout,
rename `magit-rename-branch' to `magit-branch-rename,
rename `magit-delete-branch' to `magit-branch-delete', and
rename `magit-branch-create-hook' to magit-create-branch-hook'.

Do not rename `magit-checkout', it's fine as is.  If we renamed we
would also have to also rename `magit-merge', and then we would end
up with too many symbol names beginning with `magit-branch-'.