temporarily remove ediff support, except resolve
Remove this implementation of Ediff support in Magit, to make it easier
to write a new more consistent and less "if cursor is where it is then
I am just gonna raise an error"ish implementation.

Keep `magit-interactive-resolve' because that is the most useful and
most often used part of the old Ediff support, and it also happens to
be the least problematic part.  It is less problematic than the rest
because it doesn't depend unrelated code to prepare data on its behalf
like the rest does.