no longer store current diff range in sections
The "diff range" of each diff can be unambiguously determined when the
information is actually required.  For diff buffers the range can be
extracted from the buffer-local `magit-refresh-args' and for diffs in
the status buffer (staged and unstaged) the "section context" can be
used to do so.

So we shouldn't store that information in each hunk section before it
is even required.  Also we shouldn't loose information in the process.

This commit removes the code that did store the information but does not
replace it with code to get the same (well better) information when it
is actually required.  That isn't necessary yet, because Ediff the only
feature which actually made use of the static information has been
removed in the previous two commits.

Dynamically determining the diff range will be implemented when Ediff
support is being reimplemented.