magit-wash-diffstat: support binary files
Binary files are displayed with "Bin XXX -> YYY bytes" instead of the
change count and graph. The numbers correspond to the old and new sizes
of the binary file, not lines added/removed, hence the capture groups
for these are left empty and this text left with no special faces.