magit-process-mode: don't turn on view-mode
This was a leftover from the old implementation, which didn't use a
major mode derived from `magit-mode'.  Back then `view-mode' was used so
that we could just use `view-mode-exit' instead of having to implement,
something similar ourselves.  Likewise that took care of making the
buffer read-only.

Now that we actually do use a mode derived `magit-mode' we also get that
for free.  So the use of `view-mode' no longer any benefits, and at the
same time has at least the drawback that we loose some key bindings
which are potentially useful.

Close #1259.