magit-git-repo-p: remove function
This function was problematic because it was very similar to some
related functions but at the same time rather different.  It also
sounded like it did the same as `git rev-parse --resolve-git-dir'
but actually did something very different.  It was also used only
once, so to avoid confusion, remove it.