magit-get-current-branch: return unambiguous shortname
Previously we got the full ref and then manually removed the
"refs/heads/" prefix.  Now we pass `--short' to `git symbolic-ref',
just like `magit-get-refname' used to.  Remove that now.  Also remove
`magit-get-ref' which got the full ref and was only used in the old
version of `magit-get-current-branch' so that it could do it wrong.

This function now returns "head/master" instead of "master", iff
e.g. refs/tags/master also exists.  That's a good thing, it's what
git would do.