magit-show-level-{1,2,3,4}{,-all}: define using defun
Remove the macros `magit-define-level-shower-1{,-all}' that were
previously used to define these commands.  That makes the code 14 lines
longer, oh the horror.  On the other hand this allows using handcrafted
doc-strings that are better than the generated crap used before.  There
are times when a define macro makes sense, here it was just an
unnecessary complication that saved us a handful of lines.