bring auto-reverting of file-visiting buffers under control
Do not rely on `auto-revert-mode'.  Instead add a new global minor mode
`magit-auto-revert-mode' which is explicitly called by `magit-refresh'
and only reverts buffers that visit files that are tracked in the
current repository.

Buffers are no longer reverted if they change for any other reason than
running a Magit command.  Doing that was a bad idea because users where
not aware of this happening and those that were aware might not want it.

Users who do want the old behavior back can instead (or additionally)
turn on `global-auto-revert-mode'.