magit.el: do not require iswitchb at compile-time
Instead declare variable `iswitchb-temp-buflist' as function
`iswitchb-read-buffer'.  That function is used by
`magit-iswitchb-completing-read' which also requires `iswitchb'
at runtime.

Generally it is better to require such features which might not be used
at runtime at least at compile time.  We are not doing that here because
in version 24.4 of Emacs, `iswitchb' is obsolete.  If we did require it
at compile time, then that would result in a warning.