magit-mode-hook: add magit-load-config-extensions
Previously users had to `magit-load-config-extensions' to
`magit-mode-hook' *and* then also enable the extensions they want to
use.  Now they just have to do the latter.  If this should ever lead to
an extension being loaded that the user did not want to be loaded then
that isn't a big problem.  The only extension for which that might cause
problems is `wip-save' which `magit-load-config-extensions' explicitly
ignores (because it applies to file-visiting buffers, not Magit
buffers).