add magit-with-editor and magit-with-git-editor
Split `magit-with-emacsclient' into two new macros `magit-with-editor'
and `magit-with-git-editor'.  For now only use the latter, which still
sets $GIT_EDITOR by default.  `magit-with-editor' sets $EDITOR instead
and can be used by other packages that need to use the emacsclient.
To make that easier these macros and related things will be split into
a new library.