magit-server-visit: use for commit and rebase
Generalize `magit-server-visit' and use it when rebasing, not only when
committing.  Add a function and a variable `magit-server-visit-args'.
Together this variable and these two functions handle all of the
"asynchronous argument passing logic" related to the use of the
Emacsclient.  And yes, I normally would not do that but here I
see no other way.  Remove `magit-commit-amending-alist'.