with-editor: new mode
Define a new minor mode `with-editor-mode' and the commands
`with-editor-finish' and `with-editor-cancel' which are used to end the
current buffers edit session.

Previously similar commands were defined for both `git-commit-mode' and
`git-rebase-mode'.  The "finish" and "cancel" commands for these two
modes should have been almost identical, however the "rebase" variants
were very insufficient and failed to cover most of the special cases the
"commit" variants knew about.  The "commit" variants on the hand also
weren't properly designed upfront and as a result were rather magical.

Instead of improving these to variants independently, this commit
generalizes the concept, adds hooks to adjust the generalized functions
to the needs of its users, uses the generalized mode and commands for
committing and rebasing, and defines the hook functions necessary to
make the generalized implementation work for these two users.